requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
streamlit
pytest
watchdog